+++ app/src/main/java/com/example/cocktail_dakk/ui/menu_detail/detailService/DetailRequest.kt
@@ -1,0 +1,7 @@
+package com.example.cocktail_dakk.ui.menu_detail.detailService
+
+data class DetailRequest(
+    val cocktailInfoId: Int,
+    val deviceNum: String,
+    val starPoint: Int
+)
