+++ app/src/main/java/com/example/cocktail_dakk/ui/menu_detail/detailService/DetailView.kt
@@ -7,3 +7,10 @@ interface DetailView{
     fun onDetailFailure(code:Int, message:String)
 }
 
+interface RatingView{
+    fun onRatingLoading()
+    fun onRatingSuccess(result: ratingResponse)
+    fun onRatingFailure(code:Int, message:String)
+}
+
+
